tinned_django/requirements.txt
# Always pin your requirements
# http://nvie.com/posts/pin-your-packages/
django==1.5.1
MySQL-python==1.2.3
Pillow==2.0.0

# Must-have packages (opinionated)
django-configurations==0.3
south==0.8.1
django-celery==3.0.11
celery==3.0.11
django-admin-tools==0.5.1
raven>=3.3.11,<3.4
django-model-utils==1.4.0
django-annoying==0.7.7
django-braces==1.0.0
django-guardian==1.1.1
django-widget-tweaks==1.3
django-debug-toolbar==0.9.4
django-mptt==0.5.5
django-redis-cache==0.10.0
redis==2.7.6
BeautifulSoup<4.0
django-compressor==1.3
django-ckeditor-fc==1.0
pytils==0.2.3
pymorphy==0.5.6
sorl-thumbnail==11.01
django-sortable-fc==1.0

# Tests
mock==1.0.0
factory-boy==1.2.0
django-nose==1.1
coverage==3.5.3
pinocchio==0.3.1